requirements.txt
annotated-types==0.6.0
anyio==4.2.0
bcrypt==4.1.2
certifi==2023.11.17
click==8.1.7
colorama==0.4.6
dnspython==2.4.2
ecdsa==0.18.0
email-validator==2.1.0.post1
fastapi==0.109.0
greenlet==3.0.3
h11==0.14.0
httpcore==1.0.2
httptools==0.6.1
httpx==0.26.0
idna==3.6
itsdangerous==2.1.2
Jinja2==3.1.3
MarkupSafe==2.1.3
orjson==3.9.10
passlib==1.7.4
psycopg2==2.9.9
pyasn1==0.5.1
pydantic==2.5.3
pydantic-extra-types==2.4.1
pydantic-settings==2.1.0
pydantic_core==2.14.6
python-dotenv==1.0.0
python-jose==3.3.0
python-multipart==0.0.6
PyYAML==6.0.1
rsa==4.9
six==1.16.0
sniffio==1.3.0
SQLAlchemy==2.0.25
starlette==0.35.1
typing_extensions==4.9.0
ujson==5.9.0
uvicorn==0.26.0
watchfiles==0.21.0
websockets==12.0
